+++ ekrano_tests/tests/known_issues.rs
@@ -18,7 +18,7 @@ mod submission;
 use ekrano::{
     AaConfig, Scene,
     kurbo::{Affine, Rect, Triangle},
-    peniko::{Color, ColorStop, Extend, Gradient, ImageFormat, ImageQuality, Mix, color::palette},
+    peniko::{Extend, ImageFormat, ImageQuality, Mix, color::palette},
 };
 use ekrano_tests::{TestParams, shared_test_device, smoke_snapshot_test_sync, snapshot_test_sync};
 use scenes::ImageCache;
@@ -206,34 +206,6 @@ fn test_data_image_roundtrip_extend_repeat() {
     scene.draw_image(&image, Affine::IDENTITY);
     let mut params = TestParams::new("data_image_roundtrip", image.image.width, image.image.height);
     params.anti_aliasing = AaConfig::Area;
-    smoke_snapshot_test_sync(scene, &params)
-        .unwrap()
-        .assert_mean_less_than(0.001);
-}
-
-/// <https://github.com/web-platform-tests/wpt/blob/18c64a74b1/html/canvas/element/fill-and-stroke-styles/2d.gradient.interpolate.coloralpha.html>
-/// See <https://github.com/linebender/vello/issues/1056>.
-fn test_gradient_color_alpha() {
-    let mut scene = Scene::new();
-    let viewport = Rect::new(0., 0., 100., 50.);
-    scene.fill(
-        ekrano::peniko::Fill::NonZero,
-        Affine::IDENTITY,
-        &Gradient::new_linear((0., 0.), (100., 0.)).with_stops([
-            ColorStop {
-                offset: 0.,
-                color: Color::from_rgba8(255, 255, 0, 0).into(),
-            },
-            ColorStop {
-                offset: 1.,
-                color: Color::from_rgba8(0, 0, 255, 255).into(),
-            },
-        ]),
-        None,
-        &viewport,
-    );
-    let mut params = TestParams::new("gradient_color_alpha", 100, 50);
-    params.base_color = Some(palette::css::WHITE);
     smoke_snapshot_test_sync(scene, &params)
         .unwrap()
         .assert_mean_less_than(0.001);
@@ -474,13 +446,6 @@ fn main() {
             Ok(())
         })
         .with_ignored_flag(true),
-    );
-    trials.push(
-        libtest_mimic::Trial::test("test_gradient_color_alpha", || {
-            test_gradient_color_alpha();
-            Ok(())
-        })
-        .with_ignored_flag(false),
     );
     trials.push(
         libtest_mimic::Trial::test("clip_blends", || {